# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lensio API
  version: 1.0.0
  description: |
    Indonesian KTP OCR API as a Service.
    Affordable, privacy-conscious, and production-grade identity data extraction.
  contact:
    name: Lensio Engineering Team
    url: https://lensio.dev
    email: api@lensio.dev
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.staging.lensio.dev
    description: Staging environment
  - url: https://api.lensio.dev
    description: Production environment
...
paths:
...
  /api/v1/usage:
    get:
      summary: Get Usage Summary
      description: Retrieves organization usage metrics and quota consumption for current billing cycle.
      operationId: getUsageSummary
      tags:
        - Usage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Organization usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummary'
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
        '403':
          $ref: '#/components/responses/InsufficientScope'
...
  /api/v1/account:
    get:
      summary: Get Account Details
      description: Retrieves current organization details, active keys count, and plan.
      operationId: getAccount
      tags:
        - Account
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account information
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization_id:
                    type: string
                    format: uuid
                  organization_name:
                    type: string
                  plan_code:
                    type: string
                  active_keys_count:
                    type: integer
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Lensio API Key (lensio_live_... or lensio_test_...)
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - invalid_api_key
                - insufficient_scope
                - rate_limit_exceeded
                - quota_exceeded
                - invalid_document
                - unsupported_document
                - ocr_failed
                - low_confidence
                - internal_error
              example: invalid_document
            message:
              type: string
              example: The uploaded document could not be processed
            request_id:
              type: string
              example: req_01JABC12345
...
    UsageSummary:
      type: object
      properties:
        total_requests:
          type: integer
          example: 450
        quota_limit:
          type: integer
          example: 1000
        quota_remaining:
          type: integer
          example: 550
        billing_cycle_reset:
          type: string
          format: date-time
  responses:
...
    InvalidAPIKey:
      description: Authentication failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: invalid_api_key
              message: Invalid or unrecognized API key
              request_id: req_01JABC12345
    InsufficientScope:
      description: Authorization failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: insufficient_scope
              message: "API key lacks required scope: ocr:write"
              request_id: req_01JABC12345
...